deleted templates, restarting API

+++ crescendo-backend/src/main/java/com/launchcode/crescendo/backend/models/Song.java
@@ -11,7 +11,7 @@ import jakarta.validation.constraints.Size;
 public class Song {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @NotBlank(message = "Title is required")
     @Size(min = 3, max = 50, message = "Title must be between 3 and 50 characters")
@@ -37,11 +37,11 @@ public class Song {
     }
 
     // Getters and setters
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
